load token

+++ utils.py
@@ -36,19 +36,22 @@ def init_bert(args):
 
 
 def embed_text(sentences, device, tokenizer, model, batch_size):
-    num_samples = len(sentences)
-    token_batches = [sentences[j * batch_size:(j + 1) * batch_size] for j in range(num_samples // batch_size)] + \
-                    [sentences[(num_samples // batch_size) * batch_size:]]
-    embed_batches = []
-    for batch in tqdm(token_batches, desc='tokenization', dynamic_ncols=True):
-        embed_batches.append(tokenizer(batch,
-                                    #    return_tensors="pt",
-                                       padding=True,
-                                       truncation=True,
-                                       max_length=512))
-    json.dump([dict(i) for i in embed_batches], open('tokenization.txt', 'w'))
+    if not os.path.exists('tokenization.txt'):
+        num_samples = len(sentences)
+        token_batches = [sentences[j * batch_size:(j + 1) * batch_size] for j in range(num_samples // batch_size)] + \
+                        [sentences[(num_samples // batch_size) * batch_size:]]
+        embed_batches = []
+        for batch in tqdm(token_batches, desc='tokenization', dynamic_ncols=True):
+            embed_batches.append(tokenizer(batch,
+                                           padding=True,
+                                           truncation=True,
+                                           max_length=512))
+        json.dump([dict(i) for i in embed_batches], open('tokenization.txt', 'w'))
+    else:
+        embed_batches = json.load(open('tokenization.txt', 'r'))
+
     del tokenizer
-    embed_batches = [{i: torch.Tensor(j).to(device) for i, j in z.items()} for z in embed_batches]
+    embed_batches = [{i: torch.LongTensor(j).to(device) for i, j in z.items()} for z in embed_batches]
     torch.cuda.empty_cache()
     with torch.no_grad():
         outputs = torch.cat([model(**batch).pooler_output for batch in tqdm(embed_batches, desc='embedding', dynamic_ncols=True)])
